pyproject.toml
[project]
name = "embeau-api"
version = "0.1.0"
description = "EMBEAU - Color Psychology Research Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database (PostgreSQL)
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # HTTP Client (for external API calls)
    "httpx>=0.28.0",

    # PyTorch & Computer Vision (for local models)
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "opencv-python>=4.9.0",
    "deepface>=0.0.93",

    # AI/ML Integration
    "openai>=1.55.0",
    "langchain>=0.3.9",
    "langchain-openai>=0.2.10",
    "langchain-community>=0.3.9",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=3.3.0",
    "google-generativeai>=0.8.0",

    # PDF Generation
    "reportlab>=4.2.0",
    "pypdf>=5.1.0",

    # Utilities
    "python-dateutil>=2.9.0",
    "pillow>=11.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/embeau_api"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
